requirements.txt
fpdf2
qrcode
SQLAlchemy==2.0.45
Flask==2.3.0
Flask-Login==0.6.2
python-dotenv==1.0.0
Werkzeug==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Mail==0.9.1
email_validator==2.0.0
Flask-WTF==1.1.1
Flask-Cors==4.0.0
Pillow==10.0.0